Add rollback toggle, improve image handling
Pixel color detection for the signal dialog has been improved and ported
to all other scenarios, including the camera button, rollback button,
"3 dots" to open the signal's side menu, and detecting whether the user
is already in a camera mode. It now checks for white pixels
(button text) when determining if a dialog or button is present to
reduce false positives.

win32gui.GetWindowRect has been changed to GetClientRect to be more
compatible when a user's window borders or title bar differ.

Switched to mss from PIL for screengrabs.

Rollback settings can be toggled now (default R).

An audible tone has been added to warn the user that the script is still
enabled when pressing "/" (chat), or pressing "`" or "'" (command bar).